refine cartcontext tests
