+++ frontend/src/context/CartContext.test.tsx
@@ -1,100 +1,64 @@
-import { act, renderHook, type RenderHookResult } from "@testing-library/react";
+import { act, renderHook } from "@testing-library/react";
 import { expect, vi } from "vitest";
 
-import { CartProvider, useCart } from "./CartContext";
-import { getLocalData, setLocalData } from "./utils/localStorage";
+import { fullProductDetails, preexistingCart } from "@/lib/test/fixtures/cart";
+
+import { useCart } from "./CartContext";
+import { setLocalData } from "./utils/localStorage";
+import { renderComponentWithCart } from "./utils/renderComponentWithCart";
 
 vi.mock("./utils/localStorage", () => ({
   getLocalData: vi.fn(),
   setLocalData: vi.fn(),
 }));
 
-/**
- * Helper function to render the CartProvider with a specific initial cart state
- * @param initialCartValue - The initial cart value to use in the context
- * @default initialCartValue { nw: {}, pns: {}, wls: {} }
- */
-function renderComponentWithCart(
-  initialCartValue: Cart | null = { nw: {}, pns: {}, wls: {} },
-): RenderHookResult<ReturnType<typeof useCart>, unknown> {
-  vi.mocked(getLocalData).mockReturnValue(initialCartValue);
-
-  const wrapper = ({ children }: { children: React.ReactNode }) => (
-    <CartProvider>{children}</CartProvider>
-  );
-
-  return renderHook(() => useCart(), { wrapper });
-}
-
-const fullProductDetails = {
-  id: "fake_id",
-  title: "Product Name 500g",
-  image: "https://example.com/product/p1.jpg",
-  productPageUrl: "https://example.com/product/p1",
-  price: {
-    value: "4.29",
-    per: "ea",
-    unitPrice: "0.86",
-    unit: "100g",
-  },
-  promo: null,
-} as Product;
-
-const existingCart = {
-  nw: {
-    nw1: {
-      product: fullProductDetails,
-      quantity: 5,
-    },
-  },
-  pns: {
-    pns1: {
-      product: fullProductDetails,
-      quantity: 12,
-    },
-    pns2: {
-      product: fullProductDetails,
-      quantity: 1,
-    },
-  },
-  wls: {},
-};
-
-// ----- ACTIVE INTERACTIONS -----
+// ----- USER INTERACTIONS -----
 
 describe("Given a user has an empty cart", () => {
   describe("When they add an item from a specific supermarket", () => {
-    const testProduct = {
-      ...fullProductDetails,
-      id: "test_product_id",
-      title: "Shampoo",
-    };
-
     it("Then the cart contains that item under the supermarket they selected", () => {
+      const testProduct = {
+        ...fullProductDetails,
+        id: "test_product_1",
+        title: "Shampoo",
+      };
+
       const { result } = renderComponentWithCart();
       act(() => {
         result.current.addCartItem("wls", testProduct);
       });
 
-      const itemInCart = result.current.cart.wls.test_product_id;
+      const itemInCart = result.current.cart.wls.test_product_1;
 
       expect(itemInCart).toBeDefined();
       expect(itemInCart.product.title).toBe("Shampoo");
     });
 
     it("Then the quantity of the new item is set to 1", () => {
+      const testProduct = {
+        ...fullProductDetails,
+        id: "test_product_2",
+        title: "Cake",
+      };
+
       const { result } = renderComponentWithCart();
       act(() => {
         result.current.addCartItem("pns", testProduct);
       });
 
-      const itemInCart = result.current.cart.pns.test_product_id;
+      const itemInCart = result.current.cart.pns.test_product_2;
 
       expect(itemInCart).toBeDefined();
       expect(itemInCart.quantity).toBe(1);
     });
 
     it("Then there is only that item in the cart", () => {
+      const testProduct = {
+        ...fullProductDetails,
+        id: "test_product_3",
+        title: "Flour",
+      };
+
       const { result } = renderComponentWithCart();
       act(() => {
         result.current.addCartItem("nw", testProduct);
@@ -102,7 +66,7 @@ describe("Given a user has an empty cart", () => {
 
       expect(result.current.cart).toEqual({
         nw: {
-          test_product_id: {
+          test_product_3: {
             product: testProduct,
             quantity: 1,
           },
@@ -117,7 +81,7 @@ describe("Given a user has an empty cart", () => {
 describe("Given a user already has items in their cart", () => {
   describe("When they clear their cart", () => {
     it("Then all supermarkets show no items", () => {
-      const { result } = renderComponentWithCart(existingCart);
+      const { result } = renderComponentWithCart(preexistingCart);
       act(() => {
         result.current.clearCart();
       });
@@ -132,7 +96,7 @@ describe("Given a user already has items in their cart", () => {
 
   describe("When they add a new item to their cart", () => {
     it("Then the item is added to the correct supermarket", () => {
-      const { result } = renderComponentWithCart(existingCart);
+      const { result } = renderComponentWithCart(preexistingCart);
       const testProduct = {
         ...fullProductDetails,
         id: "new_item_1",
@@ -150,7 +114,7 @@ describe("Given a user already has items in their cart", () => {
     });
 
     it("Then the quantity of the new item is set to 1", () => {
-      const { result } = renderComponentWithCart(existingCart);
+      const { result } = renderComponentWithCart(preexistingCart);
       const testProduct = {
         ...fullProductDetails,
         id: "new_item_2",
@@ -197,8 +161,8 @@ describe("Given a user has not been to the site before", () => {
 describe("Given a user has previously been to the site", () => {
   describe("When they load up the site", () => {
     it("Then their cart is restored to its previous state", () => {
-      const { result } = renderComponentWithCart(existingCart);
-      expect(result.current.cart).toEqual(existingCart);
+      const { result } = renderComponentWithCart(preexistingCart);
+      expect(result.current.cart).toEqual(preexistingCart);
     });
   });
 });
